extract label logic to new file
